Implemented delete (need to check the return)

+++ ContactList/src/main/java/UXStudio/contactList/ContactList/Repository/Repository.java
@@ -4,7 +4,6 @@ import UXStudio.contactList.ContactList.model.Contact;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @org.springframework.stereotype.Repository
-public interface Repository extends JpaRepository<Contact, Long> {
-
+public interface Repository extends JpaRepository<Contact, Integer> {
 
 }
